Split OMS auditing out to clear the file-size ratchet
verify-dotnet failed on the no-new-god-file gate rather than a compile
error: SettleReservations pushed OrderManagementSystem.cs to 2005 lines,
five over the 2000 limit.

The gate offers a baseline exemption, but the file was already at the
ceiling and taking it would just endorse the next five lines. Split
instead, along the seam the file already had: everything that answers
"write down what happened" -- the metadata builders, the lifecycle audit
append, and the rejection path that registers the rejected state, records
it, and returns the result -- moves to a partial-class file. None of it
routes an order or talks to a gateway. The main file keeps the submission
and cancellation lifecycle and drops to 1786 lines.

Named for the repository's existing Type.Aspect.cs convention.
