+++ docs/status/doc-health-dashboard.json
@@ -1,6 +1,6 @@
 {
   "total_files": 639,
-  "total_lines": 117333,
+  "total_lines": 117334,
   "orphaned_count": 257,
   "no_heading_count": 148,
   "stale_count": 0,
@@ -2580,7 +2580,7 @@
     },
     {
       "path": "docs/generated/repository-structure.md",
-      "line_count": 9633,
+      "line_count": 9634,
       "has_heading": true,
       "todo_count": 8,
       "last_modified_utc": "1970-01-01T00:00:00+00:00",
